requirements.txt
aiogram==2.25.1
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==23.1.0
Babel==2.9.1
boto3==1.26.126
botocore==1.29.126
certifi==2023.5.7
charset-normalizer==3.1.0
Django==4.2
django-filter==23.2
django-modeltranslation==0.18.9
django-storages==1.13.2
djangorestframework==3.14.0
environs==9.5.0
frozenlist==1.3.3
geographiclib==2.0
geopy==2.3.0
idna==3.4
jmespath==1.0.1
magic-filter==1.0.9
Markdown==3.4.3
marshmallow==3.19.0
multidict==6.0.4
packaging==23.1
Pillow==9.5.0
psycopg2-binary==2.9.6
python-dateutil==2.8.2
python-dotenv==1.0.0
python-gettext==5.0
pytz==2023.3
requests==2.30.0
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.5.0
urllib3==1.26.15
yarl==1.9.2
